feat(layout): add chat list layout & mention list layout
