+++ src/components/community/UserImage/styled.ts
@@ -1,6 +1,6 @@
 import styled from '@emotion/native';
 
-export const CommunityChatImage = styled.Image`
+export const CommunityUserImage = styled.Image`
   border-radius: 100px;
   width: 40px;
   height: 40px;
