+++ src/layouts/community/chat/ChatList/styled.ts
@@ -1,0 +1,16 @@
+import styled from '@emotion/native';
+
+export const ChatListContainer = styled.View`
+  flex: 1;
+  row-gap: 24px;
+  justify-content: flex-start;
+  align-items: flex-start;
+  width: 100%;
+`;
+
+export const ChatReplyContainer = styled.View`
+  flex-direction: row;
+  column-gap: 10px;
+  align-items: center;
+  margin-bottom: 10px;
+`;
