+++ src/layouts/community/chat/MetionUserList/index.tsx
@@ -1,0 +1,43 @@
+import React from 'react';
+import { FlatList, View } from 'react-native';
+
+import { useTheme } from '@emotion/react';
+
+import { COMMUNITY_USER_LIST } from 'src/constants';
+import { CommunityUserImage, ScaleOpacity, Text } from 'src/components';
+
+import * as S from './styled';
+
+export interface MentionUserListProps {
+  onMention: (id: string, isReply?: boolean) => void;
+}
+
+export const MentionUserList: React.FC<MentionUserListProps> = ({ onMention }) => {
+  const theme = useTheme();
+
+  return (
+    <FlatList
+      data={COMMUNITY_USER_LIST}
+      keyExtractor={(_, index) => index.toString()}
+      contentContainerStyle={{
+        width: '100%',
+        alignItems: 'flex-start',
+        padding: 14,
+        rowGap: 24,
+      }}
+      renderItem={({ item: { name, image, id } }) => (
+        <ScaleOpacity onPress={() => onMention(id)}>
+          <S.MentionUserContainer>
+            <CommunityUserImage userImage={image} />
+            <View>
+              <Text size={16}>{name}</Text>
+              <Text size={14} color={theme.placeholder}>
+                {id}
+              </Text>
+            </View>
+          </S.MentionUserContainer>
+        </ScaleOpacity>
+      )}
+    />
+  );
+};
