+++ src/layouts/community/chat/MetionUserList/styled.ts
@@ -1,0 +1,8 @@
+import styled from '@emotion/native';
+
+export const MentionUserContainer = styled.View`
+  flex-direction: row;
+  align-items: center;
+  column-gap: 10px;
+  justify-content: center;
+`;
