+++ src/layouts/community/chat/index.ts
@@ -1,1 +1,3 @@
 export * from './BottomInput';
+export * from './MetionUserList';
+export * from './ChatList';
